kalibrr/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
async-generator==1.10; python_version >= '3.5'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
beautifulsoup4==4.11.1; python_version >= '3.6'
bs4==0.0.1
cachetools==5.2.0; python_version ~= '3.7'
certifi==2022.5.18.1; python_version >= '3.6'
cffi==1.15.0; os_name == 'nt' and implementation_name != 'pypy'
charset-normalizer==2.0.12; python_version >= '3.5'
cryptography==37.0.2
google-api-core[grpc]==2.7.0
google-auth==2.7.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-cloud-bigquery-storage==2.13.2; python_version >= '3.6'
google-cloud-bigquery==3.2.0
google-cloud-core==2.3.1; python_version >= '3.6'
google-cloud-storage==2.4.0
google-crc32c==1.3.0; python_version >= '3.6'
google-resumable-media==2.3.3; python_version >= '3.6'
googleapis-common-protos==1.56.2; python_version >= '3.6'
grpcio-status==1.46.3
grpcio==1.46.3
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
keyring==23.6.0
lxml==4.9.0
numpy==1.22.4; python_version >= '3.10'
outcome==1.1.0; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pandas==1.4.2
proto-plus==1.20.5; python_version >= '3.6'
protobuf==3.20.1; python_version >= '3.7'
pyarrow==8.0.0; python_version >= '3.7'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.9; python_full_version >= '3.6.8'
pysocks==1.7.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0; python_version >= '3.5'
pytz==2022.1
pywin32-ctypes==0.2.0; sys_platform == 'win32'
requests==2.28.0; python_version >= '3.7' and python_version < '4'
rsa==4.8; python_version >= '3.6'
selenium==4.2.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
soupsieve==2.3.2.post1; python_version >= '3.6'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.21.0; python_version >= '3.7'
urllib3[secure,socks]==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
webdriver-manager==3.7.0
wget==3.2
wsproto==1.1.0; python_version >= '3.7'
